Project/requirements.txt
streamlitpip
pandas
secure-smtplib
numpy